maxDeduplication improved
