+++ ledger/sandbox-on-x/src/app/resources/application.conf
@@ -140,4 +140,5 @@ ledger {
 bridge {
     conflict-checking-enabled = true
     submission-buffer-size = 500
+    max-deduplication-duration = "1800 seconds"
 }
