+++ ledger/sandbox-on-x/src/test/app/resources/application.conf
@@ -129,4 +129,5 @@ ledger {
 bridge {
   conflict-checking-enabled = true
   submission-buffer-size = 500
+  max-deduplication-duration = "1800 seconds"
 }
